hotfix Docker-Java Client unboxToInt Exception

+++ streamx-common/src/main/scala/com/streamxhub/streamx/common/conf/CommonConfig.scala
@@ -48,20 +48,20 @@ object CommonConfig {
 
   val DOCKER_MAX_CONNECTIONS: ConfigOption = ConfigOption(
     key = "streamx.docker.http-client.max-connections",
-    defaultValue = 100L,
-    classType = classOf[java.lang.Long],
+    defaultValue = 100,
+    classType = classOf[java.lang.Integer],
     description = "instantiating max connections for DockerHttpClient")
 
   val DOCKER_CONNECTION_TIMEOUT_SEC: ConfigOption = ConfigOption(
     key = "streamx.docker.http-client.connection-timeout-sec",
-    defaultValue = 100L,
-    classType = classOf[java.lang.Long],
+    defaultValue = 100,
+    classType = classOf[java.lang.Integer],
     description = "instantiating connection timeout for DockerHttpClient")
 
   val DOCKER_RESPONSE_TIMEOUT_SEC: ConfigOption = ConfigOption(
     key = "streamx.docker.http-client.response-timeout-sec",
-    defaultValue = 120L,
-    classType = classOf[java.lang.Long],
+    defaultValue = 120,
+    classType = classOf[java.lang.Integer],
     description = "instantiating connection timeout for DockerHttpClient")
 
   val MAVEN_REMOTE_URL: ConfigOption = ConfigOption(
